+++ HybridHostExample-Harmony/HBuilder-uniPluginDemo/uni_modules/GC-UniPlugin/js_sdk/index.js
@@ -19,8 +19,8 @@ export {
 	gcRequest
 } from './Request/GCRequest.js';
 export {
-	gcHarmonyNetworkTracking
-} from './Request/GCHarmonyNetworkTracking.js';
+	gcResourceTracking
+} from './Request/GCResourceTracking.js';
 export {
 	gcPageMixin
 } from './View/GCPageMixin.js';